HYPEROPT: Migrate Sampler functionality to Executor (#2165)
* WIP: refactor to remove hyperopt sampler

Migrating sampler functionality to Executor

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* test: update tests for removal of Sampler

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* refactor: remove HyperoptExecutor class

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* code: remove no longer needed code

* code: remove old code

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* code: re-introduce function ray_resource_allocation_function()

* code: fix formatting

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* code: add missing import

* refactor: incorporate reviewer comment

consolidated code to simplify setup of search algorithm in executor

* refactor: incorporate reviewer comments

Remove redundant code and add typing

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* code: remove unused imports

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci
